requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.3
qrcode==7.4.2
Pillow==10.0.1
Werkzeug==2.2.3
python-dotenv==1.0.0
PyMySQL==1.1.0